revert head to previous

+++ Autopilot/Test/Src/PathManager/Test_CruisingState.cpp
@@ -4,12 +4,10 @@
 
 #include <gtest/gtest.h>
 
-#include <cstdint>
-
 #include "pathManager.hpp"
 #include "pathStateClasses.hpp"
 #include "waypointManager.hpp"
-#include "cruisingFlight.hpp"
+#include "cruisingState.hpp"
 #include "pathDatatypes.hpp"
 
 using namespace std; 
@@ -50,7 +48,7 @@ enum _ReturnToGroundStatus {RETURN_SUCCESS = 0, RETURN_DIFFERENT};
 enum _OutputStatus {OUTPUT_CORRECT = 0, OUTPUT_INCORRECT};
 enum _WaypointSt {WAYPOINT_CORRECT = 0, WAYPOINT_INCORRECT};
 
-static _ArrayStatus compare_id_arrays (uint16_t * toCheck, uint16_t * ans) {
+static _ArrayStatus compare_id_arrays (int * toCheck, int * ans) {
     for (int i = 0; i < PATH_BUFFER_SIZE; i++) {
         if (toCheck[i] != ans[i]) {
             // cout << i << " | " << toCheck[i] << " ; ans: " << ans[i] << endl;
@@ -173,7 +171,7 @@ TEST (CruisingState, IncorrectTelemetryCommandsReturnErrorCode) {
 	/********************DEPENDENCIES*******************/	
 	/********************STEPTHROUGH********************/
 
-    uint16_t idArray[PATH_BUFFER_SIZE];
+    int idArray[PATH_BUFFER_SIZE];
     WaypointManager cruisingStateManager;
     bool goingHome = false, inHold = false; 
 
@@ -218,7 +216,7 @@ TEST (CruisingState, InitializeFlightPathSuccess) {
     TelemetryTestData.homebase = createTelemetryWaypoint(0.0, 0.0, 100, 0.0, 0);
 
     // Set up id test array (desired result)
-    uint16_t idTestArray[PATH_BUFFER_SIZE];
+    int idTestArray[PATH_BUFFER_SIZE];
     idTestArray[0] = 1;
     idTestArray[1] = 2;
     idTestArray[2] = 3;
@@ -251,7 +249,7 @@ TEST (CruisingState, InitializeFlightPathSuccess) {
     _PathData * testHomeBase = testWaypointManager.initialize_waypoint(0.0, 0.0, 100, PATH_FOLLOW);
     
     // Variables that will be modified as we go
-    uint16_t idArray[PATH_BUFFER_SIZE] = { 0 };
+    int idArray[PATH_BUFFER_SIZE] = { 0 };
     bool goingHome = false, inHold = false; 
 
 	/********************DEPENDENCIES*******************/	
@@ -313,7 +311,7 @@ TEST (CruisingState, NukeFlightPathSuccess) {
     TelemetryTestData.homebase = createTelemetryWaypoint(0.0, 0.0, 100, 0.0, 0);
 
     // Set up id test array (desired result)
-    uint16_t idTestArray[PATH_BUFFER_SIZE];
+    int idTestArray[PATH_BUFFER_SIZE];
 
     // Set up flight path array (desired result)
 
@@ -328,7 +326,7 @@ TEST (CruisingState, NukeFlightPathSuccess) {
     }
     
     // Variables that will be modified as we go
-    uint16_t idArray[PATH_BUFFER_SIZE] = { 0 };
+    int idArray[PATH_BUFFER_SIZE] = { 0 };
     bool goingHome = false, inHold = false; 
 
 	/********************DEPENDENCIES*******************/	
@@ -387,7 +385,7 @@ TEST (CruisingState, AppendWaypointSuccess) {
     TelemetryTestData.homebase = createTelemetryWaypoint(0.0, 0.0, 100, 0.0, 0);
 
     // Set up id test array (desired result)
-    uint16_t idTestArray[PATH_BUFFER_SIZE] = { 0 };
+    int idTestArray[PATH_BUFFER_SIZE] = { 0 };
     idTestArray[0] = 1;
     idTestArray[1] = 2;
     idTestArray[2] = 3;
@@ -418,7 +416,7 @@ TEST (CruisingState, AppendWaypointSuccess) {
     }
     
     // Variables that will be modified as we go
-    uint16_t idArray[PATH_BUFFER_SIZE] = {0};
+    int idArray[PATH_BUFFER_SIZE] = {0};
     bool goingHome = false, inHold = false; 
 
 	/********************DEPENDENCIES*******************/	
@@ -483,7 +481,7 @@ TEST (CruisingState, AppendWaypointFail) {
     TelemetryTestData.homebase = createTelemetryWaypoint(0.0, 0.0, 100, 0.0, 0);
 
     // Set up id test array (desired result)
-    uint16_t idTestArray[PATH_BUFFER_SIZE] = { 0 };
+    int idTestArray[PATH_BUFFER_SIZE] = { 0 };
     idTestArray[0] = 1;
     idTestArray[1] = 2;
     idTestArray[2] = 3;
@@ -511,7 +509,7 @@ TEST (CruisingState, AppendWaypointFail) {
     }
     
     // Variables that will be modified as we go
-    uint16_t idArray[PATH_BUFFER_SIZE] = { 0 };
+    int idArray[PATH_BUFFER_SIZE] = { 0 };
     bool goingHome = false, inHold = false; 
 
 	/********************DEPENDENCIES*******************/	
@@ -576,7 +574,7 @@ TEST (CruisingState, InsertWaypointSuccess) {
     TelemetryTestData.homebase = createTelemetryWaypoint(0.0, 0.0, 100, 0.0, 0);
 
     // Set up id test array (desired result)
-    uint16_t idTestArray[PATH_BUFFER_SIZE] = { 0 };
+    int idTestArray[PATH_BUFFER_SIZE] = { 0 };
     idTestArray[0] = 1;
     idTestArray[1] = 2;
     idTestArray[2] = 3;
@@ -607,7 +605,7 @@ TEST (CruisingState, InsertWaypointSuccess) {
     }
     
     // Variables that will be modified as we go
-    uint16_t idArray[PATH_BUFFER_SIZE] = { 0 };
+    int idArray[PATH_BUFFER_SIZE] = { 0 };
     bool goingHome = false, inHold = false; 
 
 	/********************DEPENDENCIES*******************/	
@@ -672,7 +670,7 @@ TEST (CruisingState, InsertWaypointFail) {
     TelemetryTestData.homebase = createTelemetryWaypoint(0.0, 0.0, 100, 0.0, 0);
 
     // Set up id test array (desired result)
-    uint16_t idTestArray[PATH_BUFFER_SIZE] = { 0 };
+    int idTestArray[PATH_BUFFER_SIZE] = { 0 };
     idTestArray[0] = 1;
     idTestArray[1] = 2;
     idTestArray[2] = 3;
@@ -700,7 +698,7 @@ TEST (CruisingState, InsertWaypointFail) {
     }
     
     // Variables that will be modified as we go
-    uint16_t idArray[PATH_BUFFER_SIZE] = { 0 };
+    int idArray[PATH_BUFFER_SIZE] = { 0 };
     bool goingHome = false, inHold = false; 
 
 	/********************DEPENDENCIES*******************/	
@@ -765,7 +763,7 @@ TEST (CruisingState, UpdateWaypointSuccess) {
     TelemetryTestData.homebase = createTelemetryWaypoint(0.0, 0.0, 100, 0.0, 0);
 
     // Set up id test array (desired result)
-    uint16_t idTestArray[PATH_BUFFER_SIZE] = { 0 };
+    int idTestArray[PATH_BUFFER_SIZE] = { 0 };
     idTestArray[0] = 1;
     idTestArray[1] = 2;
     idTestArray[2] = 5;
@@ -794,7 +792,7 @@ TEST (CruisingState, UpdateWaypointSuccess) {
     }
     
     // Variables that will be modified as we go
-    uint16_t idArray[PATH_BUFFER_SIZE] = { 0 };
+    int idArray[PATH_BUFFER_SIZE] = { 0 };
     bool goingHome = false, inHold = false; 
 
 	/********************DEPENDENCIES*******************/	
@@ -859,7 +857,7 @@ TEST (CruisingState, UpdateWaypointFail) {
     TelemetryTestData.homebase = createTelemetryWaypoint(0.0, 0.0, 100, 0.0, 0);
 
     // Set up id test array (desired result)
-    uint16_t idTestArray[PATH_BUFFER_SIZE] = { 0 };
+    int idTestArray[PATH_BUFFER_SIZE] = { 0 };
     idTestArray[0] = 1;
     idTestArray[1] = 2;
     idTestArray[2] = 3;
@@ -887,7 +885,7 @@ TEST (CruisingState, UpdateWaypointFail) {
     }
     
     // Variables that will be modified as we go
-    uint16_t idArray[PATH_BUFFER_SIZE] = { 0 };
+    int idArray[PATH_BUFFER_SIZE] = { 0 };
     bool goingHome = false, inHold = false; 
 
 	/********************DEPENDENCIES*******************/	
@@ -952,7 +950,7 @@ TEST (CruisingState, DeleteWaypointSuccess) {
     TelemetryTestData.homebase = createTelemetryWaypoint(0.0, 0.0, 100, 0.0, 0);
 
     // Set up id test array (desired result)
-    uint16_t idTestArray[PATH_BUFFER_SIZE] = { 0 };
+    int idTestArray[PATH_BUFFER_SIZE] = { 0 };
     idTestArray[0] = 1;
     idTestArray[1] = 2;
     idTestArray[2] = 4;
@@ -980,7 +978,7 @@ TEST (CruisingState, DeleteWaypointSuccess) {
     }
     
     // Variables that will be modified as we go
-    uint16_t idArray[PATH_BUFFER_SIZE] = { 0 };
+    int idArray[PATH_BUFFER_SIZE] = { 0 };
     bool goingHome = false, inHold = false; 
 
 	/********************DEPENDENCIES*******************/	
@@ -1043,7 +1041,7 @@ TEST (CruisingState, DeleteWaypointFail) {
     TelemetryTestData.homebase = createTelemetryWaypoint(0.0, 0.0, 100, 0.0, 0);
 
     // Set up id test array (desired result)
-    uint16_t idTestArray[PATH_BUFFER_SIZE] = { 0 };
+    int idTestArray[PATH_BUFFER_SIZE] = { 0 };
     idTestArray[0] = 1;
     idTestArray[1] = 2;
     idTestArray[2] = 3;
@@ -1071,7 +1069,7 @@ TEST (CruisingState, DeleteWaypointFail) {
     }
     
     // Variables that will be modified as we go
-    uint16_t idArray[PATH_BUFFER_SIZE] = { 0 };
+    int idArray[PATH_BUFFER_SIZE] = { 0 };
     bool goingHome = false, inHold = false; 
 
 	/********************DEPENDENCIES*******************/	
@@ -1134,7 +1132,7 @@ TEST (CruisingState, NextDirectionsRegularCorrect) {
     TelemetryTestData.homebase = createTelemetryWaypoint(0.0, 0.0, 100, 0.0, 0);    
       
     // Variables that will be modified as we go
-    uint16_t idArray[PATH_BUFFER_SIZE] = { 0 };
+    int idArray[PATH_BUFFER_SIZE] = { 0 };
     bool goingHome = false, inHold = false; 
 
     // Set up input data for waypoint manager:
@@ -1207,7 +1205,7 @@ TEST (CruisingState, NextDirectionsGoingHomeCorrect) {
     TelemetryTestData.homebase = createTelemetryWaypoint(-80.537331184, 43.467998128, 45, 0.0, 2);
     
     // Variables that will be modified as we go
-    uint16_t idArray[PATH_BUFFER_SIZE] = { 0 };
+    int idArray[PATH_BUFFER_SIZE] = { 0 };
     bool goingHome = false, inHold = false; 
 
     // Set up input data for waypoint manager:
@@ -1295,7 +1293,7 @@ TEST (CruisingState, GoingHomeHomeBaseUndefined) {
     TelemetryTestData.homebase = createTelemetryWaypoint(-80.537331184, 43.467998128, 45, 0.0, 2);
     
     // Variables that will be modified as we go
-    uint16_t idArray[PATH_BUFFER_SIZE] = { 0 };
+    int idArray[PATH_BUFFER_SIZE] = { 0 };
     bool goingHome = false, inHold = false; 
 
     // Set up input data for waypoint manager:
@@ -1368,7 +1366,7 @@ TEST (CruisingState, NextDirectionsStartHoldingCorrect) {
     TelemetryTestData.homebase = createTelemetryWaypoint(-80.537331184, 43.467998128, 45, 0.0, 2);
     
     // Variables that will be modified as we go
-    uint16_t idArray[PATH_BUFFER_SIZE] = { 0 };
+    int idArray[PATH_BUFFER_SIZE] = { 0 };
     bool goingHome = false, inHold = false; 
 
     // Set up input data for waypoint manager:
